Implement get_numeric_arg for flags that need arg
e.g., "todo -a" without a numeric arg following the flag was segfaulting
because we weren't checking the value of argc. Now we wrap atoi_pedantic
in this helper function (couldn't just check in main() because the
fallback behaviour is to accept any non-flag args as a message, so it
would have required a lot of extra lines of code).