Implement major state model changes and refactor
Uses an enum to track requested input option, a function to extract
option state from an option string, and a switch in the main loop to
choose the appropriate action based on the enum. Deprecates *all* of
what was previously in main(), which was hacky anyway.

Line printing and line marking refactored to use this state enum as an
input flag to choose appropriate behaviour. mark_xxxx() now completely
deprecated in favour of a generic mark_line that chooses the correct
behaviour based on its input action. print_lines minor refactoring the
same.